Reduce copying and Sliver construction/destruction in comparator
Profiling showed that a read/write-heavy workload spent significant
amount of time in the rocksdb key comparator, and that most of that
time was spent copying buffers, and constructing and destructing
Sliver objects. This MR adds versions of the functions required by the
key comparator that operate directly on pointers, instead of expecting
a Sliver wrapping. This removes copying and extraneous Sliver
construction in this path, which showed performance improvements in
read/write-heavy tests.