Avoid bogus "receiver is NULL" log messages when using UDP.
Every simpleTest run was printing "Node #: receiver is NULL" at the end of
the test logs.  This didn't really make sense because the code sets
receiver nonnull and then never sets it back to null afterward.  However,
a received message length of 0 could also cause the message to be printed,
and that was the issue.  Normally UDP would only return a 0-length message
if the sender was sending 0-length messages, but it also turns out that
(at least with Linux) using shutdown() to half-close a socket for receive
also causes recvfrom() to return 0-length messages.  The Stop() member
function does exactly that, as a way to wake up the receive thread when
it's time for it to exit.  Thus, this is a perfectly normal thing in this
case and doesn't warrant logging a (wrong) error message.  This commit
changes the code to instead log an accurate message at debug level.