Fix uninitialized variables in TestGeneratedKeys.
GCC 7.4 with -O reported that numReplicas "may" be used uninitialized
in main() in this program.  It doesn't seem to be right about it, but
initializing it silenced the warning.

At the same time, I noticed that hasNumReplicas and hasOutputPrefix do
seem to be actually used uninitialized, although GCC didn't warn about
them, so this commit also fixes that.