Make Sliver immutable and back by char[] or string
A Sliver's data is contained in a contiguous buffer wrapped in a shared
pointer. There is no mutex protecting the data, and it is unsafe to
mutate. None of our code ever mutates a Sliver anyway, so modify its API to
make it immutable.

Additionally, allow backing the Sliver by a string as well as a char[].
RocksDB returns strings in Get requests, and previously we were copying
them into a buffer and putting it into the Sliver. This change adds a
move constructor for strings and uses std::variant to allow safely
putting a string or char[] behind a shared_ptr. There is no
corresponding copy constructor for a string, as that leads to accidental
copies, when that is likely not the intent. String copies can be made at
the call site and moved in if necessary.

std::variant is only available in C++17, so bump the cmake setting to
allow it. Using C++17 opens up a bunch of other useful tools, like
std::string_view and std::make_shared. make_shared in particular reduces
allocations from 2 to one when constructing a new value directly inside
a shared_ptr, and so we use it in this commit.

The original backing type of Sliver was also changed from a uint8_t[] to
a char[], and the constructors relating to uint8_t were removed. Most C
and C++ code uses char* pointers which are much more common than
uint8_t*, so codify that in the API and backing. Additionally make them
all take const*, to indicate that a Sliver is immutable. This works in
our code because a const* parameter can take a non const argument. The
previous constructors that took const char* and const uint8_t* but did
not take ownership were removed. These constructors were unsafe and
prone to error and confusion as they did not match the lifetimes of the
data with the Sliver. Since slivers are purposefully constructed to
safely share data, this was especially dangerous.

The RocksDB client was updated to use the new string move constructor in
the Get method.

Some overly complex code was removed from memorydb client, thus
simplifying the get, put, and del methods. Note also that we use the
C++17 `insert_or_assign` map method in the put method, as that saves a
key copy if the key already exists. A bug was also fixed in the get
method that takes a char* out buf to be filled in, and an unnecessary
copy was removed. We didn't notice the problem because our tests at
least always use the exact size of the buffer and so it matches the
real size. The code now looks more like the RocksDB client
implementation and uses a logger as well.