Don't assert when sendto fails in UDP communication
sendto can fail for various reasons including a message too large and
the send buffer being full. We shouldn't kill the process when this
occurs.

This was seen in a CI test run:
 https://travis-ci.com/vmware/concord-bft/jobs/235340703#L2443

This is part of a fix when debugging #212, although it doesn't appear to
be the direct issue from the log linked above, since the test that
failed came after the test containing the assert.