State transfer persistence changes pass tests
Tests now pass for in-memory and rocksdb based state transfer
DataStores.

This commit also makes a number of changes for clarity and safety. The
InMemoryDataStore and memorydb::Client are not threadsafe. I'm not sure
the rocksdb client is either. Therefore we don't create shared_ptr
wrappers around them. While the shared pointers were used in a single
thread, it was unclear that this is the case without a broad
undertsanding of the code base. If this ever changes, it would be a
problem.

I also removed the typedefs in db_interface.hpp relating to
shared_ptr, because it makes the calling code less clear. The user would
have to look in that file to discover what type of pointer was being
used. We also don't really want to mandate shared pointers, as that
should be a user choice. Lastly, as stated above, this usage could be
unsafe since those types are not threadsafe.

This commit also removed the DataStoreTransaction wrapper, since it
relied on shared pointers around the DataStore which is not thread safe.
The DataStoreTransaction is also not currently used, and I don't believe
it's needed. Instead, the DbDataStore put methods now take an optional
ITransaction* if needed. Likely using this will require a refactoring,
since it may need to be included in other DataStore methods that take
transactions inside guards, or a transaction around those methods and
the put will be needed. Right now though, neither is needed so we just
make the possibility available for future refactorings.

Lastly, the majority of DataStore get methods were made const, since
they don't mutate state.