Various fixes for bcstatetransfer persistence
The DBDataStore was leaking copies of itself whenever a transaction was
created. This is because a new DBDataStore copy was being allocated on the
heap and used by the DataStoreTransaction, but never freed when the
transaction ended. This was fixed via instantiating the
DataStoreTransaction with a reference to the DBDataStore rather than a
new allocated copy. This is justified because a DBDataStore and its
underlying IDBClient are not threadsafe, and so borrowing a reference is
both more efficient, and less likely to lead to confusion. We also
ensure that in DataStore::beginTransaction() calls that a transaction is
not currently running to guarantee safety.

Now that we don't copy DataStores, we must ensure that the underlying
IDBClient tranasaction in the txn_ field of DBDataStore is freed and set back to
null when the transaction ends. In fact, it looks like this underlying
ITransaction was leaked as well. A new DataStore method was added,
endTransaction() that frees and resets the pointer in the DBDatastore
txn_ field. It runs when the DataStoreTransaction is destructed. A
corresponding version was added to InMemoryDataStore.

DBDataStore transactions also should only wrap database backed
DataStores, currently only rocksdb. bcstatetransfer_tests were fixed for
the in memory usecase by passing in either a DBDataStore or
InMemoryDataStore rather than always passing in a DBDataStore that
sometimes wrapped an in-memory client. This seems to have been the
intention, since the NullTransaction exists in the InMemoryDataStore
class, but not in the memorydb code for IDBClient which currently
returns a nullptr.