Fix skvbc_persistence_tests
An assert was firing on restart of replicas during
`BCStateTran::checkConsistency`: ERROR 2019-09-30 14:32:14.865 (concord) Assert: expression 'lastBlockNum >= lastReachableBlockNum' is false in function checkConsistency (/home/andrewstone/concord-bft/bftengine/src/bcstatetransfer/BCStateTran.cpp 2047)

In this case lastBlockNum = 0, lastReachableBlockNum = 1.

It turns out this was due to the way the lastBlockNum is retrieved from
the blockchain::DBAdapter:
https://github.com/vmware/concord-bft/blob/master/storage/src/blockchain_db_adapter.cpp#L890-L919.

The last block is found by seeking to the largest block and then walking
back one to the previous block. This works if the blocks are in
ascending order. However, they were mistakenly switched to be stored in
descending order by this change:
https://github.com/vmware/concord-bft/commit/be97c2d8898659795690d4c6d182f975de0fe65c#diff-d7e572309a4d8c96992ba11b40b507d3L73-L109

If they are stored in descending order, searching for the max size block
will end up with the iterator at position 0 and there will be no
previous block, thus the last block will always be 0. We were also
seeing this bug reflected periodically in performance tests:
`2019-09-19T21:40:15.589 [140094483285760] ERROR state-transfer %% ‘lastBlock == as_->getLastBlockNum()’ is NOT true (in function ‘createCheckpointDesc’ in /concord/submodules/concord-bft/bftengine/src/bcstatetransfer/BCStateTran.cpp:323)`

By properly ordering blocks in ascending order when the type is
`EDBKeyType::E_DB_KEY_TYPE_BLOCK` we have fixed this bug.

Additionally clean up the checkConsistency method to return void and
assert directly on failure rather than returning a bool which is then
asserted on. This gives us the line where the failure occurred rather
than just telling us something failed inside the method.

Also remove a few repetitive Assert macros and use the Assert macro in
assertUtils.h.

Finally, fix a bug in DBDataStore::deserializeCheckpoint so that
desc.lastBlock is deserialized along with the other Checkpoint
descriptor fields.