Allow custom aggregator when creating BCStateTran
Add an overload to bftEngine::SimpleBlockchainStateTransfer::create such
that it takes an std::shared_ptr<concordMetrics::Aggregator> as a
parameter, such that BCStateTran can then update a global aggregator
with its metrics. Set this aggregator in simpleKVBC::ReplicaImp so we
can use it in python system tests via the udp metrics server.