Split source selection from BC ST into own class
This commit is part of a refactoring of BCStateTran to make the logic
easier to understand. Ideally we separate out the low level
implementation details from the high level control in the `fetch()`
loop. Note that `processData` was changed to `fetch` which is a bit more
descriptive about what it does.

The goal was not to change behavior and all tests currently pass.

This commit also properly loads relevant metrics on initialization of
BCStateTran.

Also fix an overflow bug similar to the one
[here](https://github.com/vmware/concord-bft/pull/78).