Split source selection from BC ST into own class
This commit is part of a refactoring of BCStateTran to make the logic
easier to understand. Ideally we separate out the low level
implementation details from the high level control in the `processData()`
loop.

The goal was not to change behavior and all tests currently pass.

This commit also properly loads relevant metrics on initialization of
BCStateTran.

Also fix an overflow bug similar to the one
[here](https://github.com/vmware/concord-bft/pull/78).

Additionally, Separate the recording of time of last sent fetch message
from the time of source selection (including received item message).
This allows detecting timeouts multiple times for retransmission before
changing a source. This appears to be the intention of the original code
since there are two separate timeouts:
 * config.fetchRetransmissionTimeoutMilli,
 * config.sourceReplicaReplacementTimeoutMilli

A change was introduced in
commit 1db691abe6c166cb5b8d534787a5bba4ce487c34 that mistakenly reset
the source selection time when a retransmission timeout occurred, but
this would mistakenly mean that the source would not ever be replaced
from timeouts alone during fetching. This change instead resets the
send time when a retransmission occurs, so you get multiple attempts at
retransmission for the same source, and then it will be replaced when
the source needs to be replaced because
config.sourceReplicaReplacementTimeoutMilli expired.