BC ST: Prevent discard of valid reserved page data
When a request for missing reserved pages gets sent because a
retransmission timer has expired, we must reset the source selection
time such that we don't immediately send a follow up request. The
BCStateTran code only allows one valid outstanding request at a time,
and if we immediately send another one, any ItemDataMsgs received from
the same source will be discarded because because the sequence number
of the original request/response will be different from the last
outstanding request sequence number. This manifests in a lot of wasted
data being sent and results in a large number of repeated warning
messages like the following:

```
WARN 2019-10-05 12:16:28.915 (state-transfer)BCStateTran::onMessage(ItemDataMsg) - msg is irrelevant: state=GettingMissingResPages, replicaId=0, currentReplica_=0, m->requestMsgSeqNum=6586328107563614208, lastMsgSeqNum_=6586328107618140160, blockNumMatches=1, dataSize=2048, totalSizeOfPendingItemDataMsgs=538624, maxPendingDataFromSourceReplica_=268435456
```

As you can see the only difference is that the sequence number of the
request: `lastMsgSeqNum_`, doesn't match that of the response:
`m->requestMsgSeqNum`. Note that we can get these warnings in case of
legitimate timeouts and delayed responses, but we shouldn't get them
when everything is working properly.

This commit resets the timer with a call to
`sourceSelector_.setSourceSelectionTime(currTime);` after a request for
reserved pages has been sent that was the result of a timeout.

Additionally, this commit makes a small change to the SourceSelector
class to not pass a member variable unnecessarily as a parameter to a
private method.