Add a state transfer persistence system test
Add a test, `test_checkpoints_saved_and_transferred`, that ensures that
checkpoints (checkpoint descriptors + blocks + reserved pages), can be
saved to disk and transferred successfully to replicas that do no have
them. Success is determined via checking internal metrics as well as
reading and writing of data. See the comments in the test for a fuller
description.

Additionally:
 * Add some methods to bft_metrics.py to allow for efficient
   multi-metric retrieval
 * Add some helper methods for waiting for checkpoints to bft_tester.py
 * Move `assert_successful_put_get` from test_skvbc.py to bft_tester.py
   so that it can be used in multiple tests.