Allow hostnames/fqdns instead of IP addresses for PlainTCP and TLSTCP
This change allows the config to specify resolvable names as replica
and client hosts, instead of requiring IP addresses, for the PlainTcp
and TlsTcp modules. The Plain UDP module is left assuming that "host"
is IP-like, because changing that assumption would require a more
in-depth change (the node mapping table relies on IP addresses, and is
set up only at module construction, for example).

For the TCP and TLS implementations:

If a hostname is used for local principals, that hostname must be
resolvable at startup time. If it is not, the node will fail to start,
after throwing an exception while trying to set up its listening
sockets.

If a hostname for any remote principal is not resolvable at startup
time, it is treated as any other connection failure and retried later.