Formatting (#263)
* Configure BFT-engine library white spaces using Google-style formatter.

* Fix format-check automatic launch during a build

* Only force a format check during CI builds

Don't add format-check to the default target. This prevents formatting
on every compilation during local development when buffers may still be
open with the files changing underneath.

* Turn on verbose logging for clang-format

This is to try to diagnose CI failures better.

Additionally, add the path that format-code.sh is operating on to cmake
output.

* Install deps to /tmp instead of $TRAVIS_BUILD_DIR

This mistakenly installs them inside the concord-bft repo and causes
code formatting scripts to run over the deps.