Fixing issue #269
URL: https://github.com/vmware/concord-bft/issues/269

It turns out, the test failure is actually a pure testing problem.

State transfer is designed to fetch blocks up to the latest stable sequence number. From then on, the stale replica catches up with the others via the BFT consensus path.

In this test however, the primary is down and view change has not triggered yet as there are no new incoming requests (by default view change does not trigger automatically).

In such circumstances, to consider state transfer completed, we should wait for the block with the latest stable sequence number to be retrieved, instead of the latest executed one.