Fix leak of ReplicaImp.m_appState
Valgrind on concord turned up this warning:

==1== 267 (40 direct, 227 indirect) bytes in 1 blocks are definitely lost in loss record 438 of 976
==1==    at 0x4C3017F: operator new(unsigned long) (in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==1==    by 0x68B137: concord::consensus::ReplicaImp::ReplicaImp(concord::consensus::CommConfig&, concord::consensus::ReplicaConsensusConfig&, concord::storage::blockchain::DBAdapter*, concord::consensus::ReplicaStateSync&) (replica_imp.cpp:276)
==1==    by 0x483ED6: run_service(concord::config::ConcordConfiguration&, concord::config::ConcordConfiguration&, log4cplus::Logger&) (main.cpp:349)
==1==    by 0x48698D: main (main.cpp:547)

Once Toly's PR #276 merges, concord should use the ReplicaImp in this
repo, and not the duplicate in its own. So, the fix is submitted here.

The leak is fixed by making m_appState a unique_ptr so that we're sure
to release it when the ReplicaImp object is deleted.