Fixing intermittent failure of skvbc-persistence-tests
This PR improves the stability of the BFT system tests when persistence is enabled.

The failures were mostly observed when restarting the stale replica and right afterwards, trying to check whether state transfer has completed. To do so, we need a number of metrics from the stale replica, as well as from an up-to-date node.

With this PR we introduce retry logic in the MetricsClient.This is needed in case we send a UDP get request to the Metrics server while it isn't listening yet. In this case we might end up waiting for a reply whereas the server hasn't received our request in the first place.

We also improve slightly the test logic, in order to make it more robust.

Testing done: ran the failing test 20 consecutive times without failure (previously it would fail every 5-10 runs)