Utilities for BFT network partitioning
With this PR we introduce the bft_network_partitioning module. It offers basic utilities and abstractions for simulating "adversaries" that are capable of interfering with network communication.
The base class for such adversaries is "NetworkPartitioningAdversary". It provides several utility methods for dropping packets based on iptables, as well as an abstract interfere() method.
Several sample implementations are provided:
- PassiveAdversary - basically the "no-op" adversary
- PrimaryIsolatingAdversary - identifies and isolates the primary replica's outgoing communication
- PacketDroppingAdversary - drops random packets between all replicas

This PR also provides to samples of how the partitioning module can be used:
1) Introduce the PrimaryIsolatingAdversary into view change tests (verify that view change is triggered due to the presence of such an adversary)
2) Introduce the PacketDroppingAdversary into linearizability tests (with 5% of packets dropped). Going above 5% significantly slows down the test, even though it still succeed.

PS: using iptables requires root privileges, hence the "sudo" addition to the commands for starting system tests.
PPS: all BFT-related partitioning rules are defined in a custom iptables rule chain called "bft-network-partitioning". This rule chain is created/cleaned-up as part of the adversary object's lifecycle (via context manager methods)