Use new Timers class in ReplicaImp
This class adds type safety and reduces the amount of work required to
add a new timer.

This change replaces the use of the old Timer and SimpleOperationsScheduler with
the new Timers class.

As the new Timers class uses `std::function` for callbacks, we can capture
`this`directly and no longer need to pass it as a void* parameter and
then cast it to InternalReplicaAPI. Timers also can now be scheduled to
recur and thus remove the need to start them on every callback. And
since these timers are only ever created and manipulated in ReplicaImp,
the above 2 changes allow us to remove all the xxxTimerHandlerFunc
wrapper functions in ReplicaImp, and call the handlers directly via a
closure. It also allows us to remove the callback functions from the
InternalReplicaAPI, since they are unused.

Since these timers are never referred to outside ReplicaImp, we also
remove the accessor functions from InternalReplicaAPI. If we need access
to these specific timers from other modules in the future we can create
a more concrete mechanism to wrap the timer manager and handles.

The Timer callback signature itself was also changed to take a Handle
when a timer fires so it can be reset or cancelled from within the
callback.

A future commit will remove the old TimeUtils code from util and replace
integer usages of time with std::chrono types.