Replace int based time with std::chrono types
This removes the need for switching from milliseconds to microseconds
and vice versa manually and adds a level of type safety.

In all cases the use of absDifference appeared unnecessary, since we are
using monotonic times and we never read them from disk, so they can't go
backwards.

Future changes may want to change constants and member variables for
configured values to use chrono types as well rather than `xxxMilli`.