Disable copy operations on RSASigner and RSAVerifier
Current implementations of copy operations on RSASigner and RSAVerifier
will cause core dumps because of double free of the internal
implementation pointer. This PR does the following:
*  Disable copy operations on RSASigner and RSAVerifier
*  Allow move operations on RSASigner and RSAVerifier
*  Implement the pimpl idiom via a std::unique_ptr to avoid casts and
manual memory management

Copy operations might be added in the future if a need arises. That will
need further investigation on the safety of CryptoPP copy operations.