Test BFT engine view change
With this PR we introduce several basic view change test scenarios:
- single view change when only the primary crashes
- single view change when f replicas (including the primary) crash
- multiple view changes on the slow path

Note: view change on the fast path cannot be tested easily by bringing down the primary.
This is because when the primary goes down, we always downgrade to the slow path for 64 sequences
and the fast path is restored only afterwards.

PS: automatic view change will be tested in a separate PR and it will allow to test
fast path & view change combined