Fixing flawed fast path assertion logic
In order to assert fast path is prevalent, we only need two things:
- the total number of sequences processed on the slow path
- the number of slow paths up to a previous known point in time ("so far")

The number of sequences processed on the fast path between this last known point in time and now,
is equal to the difference between those two numbers.

The "as_of_seq_num" parameter was redundant and resulted in flawed logic.

Note: the tests used to pass because, due to this flaw, the fast path assertion logic was too "permissive".