SparseMerkle: Add BatchedInternalNode
The BatchedInternalNode is the first major piece of the sparse merkle
tree data structure. A BatchedInternalNode contains up to 4 levels of a
merkle tree, and is identified by an InternalNodeKey, used for database
access.

Insertion into a BatchedInternalNode works now, such that keys can be
added to a single node. In some cases, such as matching prefixes of a
certain length, this can cause the node to split. In this case we inform
the caller of this and tell it to create new nodes and insert the old
stored leaf that triggered the split and the new child attempting to be
inserted. In other cases, insert can discover that the leaf attempting
to be inserted doesn't fit in this node and needs to be placed further
down the tree, into an existing node. There is a return value indicating
this as well.

The logic to handle the DB access as well as creation of new nodes and
updates of hashes when required by the creation of a new node is
relegated to the sparse merkle tree logic itself which will come in
further commits.

Base types were modified in order to support the BatchedInternalNode.
Tests were also added to cover the variety of insertion scenarios.

This code heavily uses value types to reduce allocations and keep things
on the stack for performance reasons.