Add an API for metrics collection
Add a generic API for metrics collection.
This API exposes 3 methods:
1. increment - to increment a counter
2.  set - to set a gauge
3. update to update a status
each metric is identified by its metricPath.
A metrics path is a path, separated by "/" in the following format:
{replica_id}/{component_name}/{metric_name}.

For example, in order to increment the replica/receivedClientRequestMsgs metrics, the replica need to call to concordMetrics::ConcordbftMetricsCollector::instance().increment(std::to_string(config_.replicaId) + "/replica/receivedClientRequestMsgs");

In addition, we add an implementation for this API that uses the existing metric framework.

restore config files