Add a container-like read-only blockchain view
Provide a container-like type (BlockchainView) that acts as standard
containers as much as possible. It tries to mimic the interfaces of a
const std::vector and std::string_view. Rationale is that it can be used
in standard algorithms, range-based for loops, looks familiar to C++
developers, etc.

BlockchainView contains block information objects of type provided by
the user. This allows for extensibility and generality. Block
information is divided in two categories - indices and data. Control
over when and how to load it is left up to users. Additionally, block
information is automatically cached inside the BlockchainView on storage
access.

NOTE: Added boost libraries in Travis so that we can run tests for
boost-dependent code.