SparseMerkle: Add initial tree implementation (#349)
* SparseMerkle: Add initial tree implementation

The first implementation of the tree was added, such that batches of
key-value pairs can be inserted via a call to `Update`. The call to
update is an iterative wrapper around calls to
`BatchedInternalNode::insert`, with one call for each KV pair in the
update.

A call to `Tree::update` returns an `UpdateBatch` with data that should
be written to the database. Besides internal and leaf nodes of the
latest updates that need writing, the `UpdateBatch` also contains
information about stale nodes such that we can prune those nodes later.
The database is never directly modified.  This allows for easier testing
and the opportunity to batch disperate writes into single transactions.
However, in order for `update` to operate, it must be able to read
existing key-value pairs from the database. This is performed via the
`IDBReader` interface. Furthermore a cache of these nodes is created and
maintained during each update call, since multiple BatchedInternalNodes
may be modified via multiple insert calls and we always want to operate
on the latest updates.

Each insert must find the correct BatchedInternalNode to insert into,
and then walk back up the tree to the root updating the hashes along
the way. This is performed via the use of a stack. Each
BatchedInternalNode is pushed onto the stack as we descend the tree and
then popped off when we reverse course and update toward the root. This
pop loop occurs when a BatchedInternalNode::insert call returns
`InsertComplete`.

Note that all KV pairs in an update call share the same version, and
therefore, each modified BatchedInternalNode or LeafNode that needs to
be written to the database will be keyed by the same version.

There are a bunch of test covering the update behavior, and test
implementation of IDBReader, called `TestDB`.

Additionally, a bunch of the underlying types were enhanced to add new
operators for comparison, etc...

Lastly, during testing two existing bugs were found:

 1. Nibble.popBack was improperly returning the first Nibble of full
 bytes because an `or` was being used instead of an `and`. The test for
 that function hid it, because the test data was an `F`, which just so
 happens to be the data returned by the bug. The existing test was
 change to use a `7` rather than an `F`. This shows the need for
 property based testing where we can test "every possible input" (via
 randomness) and get more input coverage. We should bring in RapidCheck
 for this.

 2. In `BatchedInternalNode::insertTwoLeafChildren`, we were inserting
 at the wrong index in the node because we were calling
 `child_key.getBit` on the number of prefix_bits_in_common, when we
 should have been starting from the next bit after that to find the
 insertion points, since the two children matched up to
 prefix_bits_in_common and differ in the next bit.