Add support for a block info initialization value in BlockchainView
BlockchainView's handling of BlockInfo is lacking in terms of a missing
option to pass arbitrary data when constructing BlockInfo objects. For
example, the user might want to pass pointers to storage handles in
order to load indices and data. This PR remedies that by adding a
BlockInfoInit type to BlockchainView and calling the BlockInfo
constructor with the passed initialization value. Backward compatibility
is preserved by making BlockInfoInit optional. Tests are added to show
the new changes work.

Additionally, this PR adds more documentation about the requirements on
the BlockInfo type. Tests are added too.