Fixing the "read-your-writes" checks in view change tests (#359)
* Fixing the "read-your-writes" checks in view change tests

The key takeaway from this effort is that our current view change mechanism is actually quite slow (in the order of 10-s of seconds).
This is why we need to increase the view change timeout, to avoid a second view change kicking-in because the first one did not complete in time.

From our discussions with Guy, it appears the "sanitary minimum" for the view change timeout is around 10 seconds (it should be noted that he was testing with 20 seconds). We may also need to increase this value to make the test more stable. Additionally, to help view change complete in a single iteration, we also bump up the network communication buffer size (necessary for networks bigger than 4 nodes).

The rest of the changes presented here mainly have to do with making the wait_for_view() implementation more realistic, waiting for the view change to actually become active, instead of just agreed.

This PR introduces the following changes:
- [SBFT] Introduce a new SBFT metric for tracking the current active view of a given replica
- [SBFT/TesterReplica] Increase the buffer size for network communication when testing
- [Tests] Await the skvbc.read_your_writes() calls
- [Tests] Modify wait_for_view() so that it waits for the new view to become active (in addition to have been "agreed upon")
- [Tests] Make sure the next expected primary is not stopped (otherwise two view changes in a row occur)
- [Tests] Disable test_multiple_vc_slow_path() for now, as it is unstable (we need log4cpp in CI to further investigate)