Don't remove root BatchedInternalNode with a child
The previous behavior of BatchedInternalNodeRemove always returned
`RemoveBatchedInternalNode` when there was one remaining LeafChild, as
it was capable of being promoted to the parent node. However, when the
LeafChild lives in the root BatchedInternalNode, it cannot be moved up
any further. In this case we should return `RemoveComplete`, which this
commit does.

Additionally, do some minor cleanup related to a review of a prior
commit.