SparseMerkle: Complete key removal functionality
This commit builds upon the previous work that enabled deletion of
individual keys inside BatchedInternalNodes. This enhances the update
functionality, such that keys can be removed as well as added and
overwritten during a `Tree::update` call.

A remove operation is the inverse of an addition, and consequently, must
reverse consequences of addition like splitting nodes. For example,
given a tree with depth 2, that looks like the following, where there
are 2 keys in the bottom BatchedInternalNode, labeled with it's nibble
path, "b":

      Root
       |
      "b"

If the hashes of the two keys are "bax" and "bex", and "bax" is
removed, then the "b" node will also be removed, and "bex" will be
**promoted** to live in the root node. Furthermore, "bex" will move up the
4-level tree inside the root BatchedInternalNode such that it resides at
the highest level in that internal tree where it has a corresponding
peer node. If "bex" is the only child in the root node, then it will
live at depth 1, right under the root of the BatchedInternalNode tree.

Note that if the tree was deeper, it is possible that "bex" would be
promoted all the way to the root still. An example of this is if the two
keys were "bex" and "bey". They would then generate a
BatchedInternalNode at depth 3 named "be", as in the tree below:

     Root
      |
     "b"
      |
     "be"

Removing one of them would cause the other to move upwards until it
reached the root.

The purpose of these examples is to display the primary invariant of a
remove operation on a tree:

    After the removal of a key on a tree, the tree should look the same
    as if the key was never inserted, without regards to versions.

In other words we should maintain the shape of the tree from before a
key is added, after that key is removed.

It's important to realize, that this promotion/compaction, while
computationally more complex, is actually an optimization. Without it,
new versions of BatchedInternalNodes further down the tree would need to
be written to disk.  Additionally, proofs would require more reads from
disks given a deeper tree, but the same number of keys. We always want
to keep the BatchedInternalNodes as densely populated as possible to
reduce disk IOPS.

A key component of the algorithm is the `Walker` class, that manages
loading and caching BatchedInternalNodes as part of the tree traversal.
It also manages the tracking of stale BatchedInternalNodes.
It is used for both insertion and removal of keys. The architectural
separation is such that the Walker is supposed to walk the tree and then
the Tree code itself will mutate the current BatchedInternalNode. This
isn't 100% true, but that's how you can think about it. Additionally,
the walker tries to abstract the cache from the insert and remove
algorithms. This isn't 100% true either, as stale Leafs are still
directly marked from the Tree code.

Many tests were added that construct trees of certain shapes and remove
keys to trigger different code paths. These tests use the following
types of assertions upon the UpdateBatches returned from update
operations to help ensure correctness:
 * The correct nodes and keys are updated
 * The updated nodes are properly linked to form a subtree
 * The correct stale nodes are returned
 * The nodes contain the correct keys
 * The nodes have the proper internal shape

Additionally, other functionality was added as part of this PR:
 * `toString()` methods were added for many types to aid in debugging
 * Types such as `UpdateCache` and `Walker` were moved into their own
 files
 * Base types were enhanced with `<<` operators to obviate the need for
 string allocation in future logging calls.
 * gitignore was enhanced to ignore apollo test files