Add support for merkle tree blocks in DBAdapter
Implement functionality to add, get and manipulate blocks in the merkle
tree DBAdapter. That includes serialization of raw blocks that are
passed by users (including state transfer) and also of block nodes that
are saved under a block key type in the DB. Block nodes only contain the
set of keys in the block and raw blocks (including values) are created
on demand (e.g. not directly saved in the DB) from the merkle tree.

Functionalities added depend heavily on the key ordering as defined in
db_types.h . Examples are getting the latest block in the system and
getting the latest internal root node in the merkle tree. Notes are
added at relevant places in the code to explain the intention.

Add the last() method in IDBClientIterator to support a well-defined
mechanism of getting the last element as opposed to getting the
previous element from the end - for example, RocksDB iterators don't
provide a notion of a 'past-the-end' iterator as C++ containers do.

Test with both a memory DB client and a RocksDB client.

Note: This commit uses the <experimental/filesystem> header (in
merkleTreeAdapter_test.cpp) from the standard library if is
not available.