Add IObjectStoreClient for Read Only Replica usage
Currently the code in #394 has a concrete ObjectStoreClient inherit
from the IDBClient pure abstract class. However, the IDBClient interface
is very powerful and object stores cannot implement it in its entirety.
Therefore the implementation of ObjectStoreClient implements most
methods with the following:

      throw std::logic_error("Not implemented for ECS S3 object store");

In addition to this, object stores have different capabilities and as
such the ObjectStoreClient class has methods not inherited from the
IDBClient interface. This leads to downcasting in order to call these
methods.

Both incomplete inheritence and downcasting suggest that an
ObjectStoreClient should not inherit from IDBClient.

The rationale for doing it this way was to be able to test the
ReadOnlyReplica (specifically RorAppState) without an available object
store. We could then instead use the MemoryDBClient or RocksdbClient for
testing via the same interface.

This commit instead introduces an IObjectStoreClient interface that can
be used as the basis for interacting with any object store in a generic
manner. To satisfy the goal of being able to use the MemoryDBClient and
RocksDBCient, we implement the interface in a DbObjectStoreClient that
can be used during testing. This allows a more complete implementation of the
production ObjectStoreClient without the need for unusable methods, and
prevents the need to downcast during testing.

It is suggested that the code in #394 use the IObjectStoreClient
abstraction instead of the IDBClient abstraction to represent
ObjectStores, and to rename the currently implemented ObjectStoreClient
to `S3ObjectStoreClient` since it uses the S3 API. To maintain the
abstraction, and prevent downcasting, the instantiation of the
S3ObjectStoreClient can be done differently, and then the
IObjectStoreClient used everywhere.