Simplify key searches by introducing seekAtMost()
A typical operation on merkle tree keys is to search for a key that is
less than or equal to a given key. Current implementation uses
seekAtLeast() and additional logic to implement these kind of searches.
This PR aims to simplify code by introducing the seekAtMost() method
that can be used in the mentioned cases.