Apollo support for an external BFT network
This PR prepares the existing Apollo tests to be run as part of the Concord product CI.

The most important change here is adding support for an externally provided BFT network. Instead of initializing the BftTestNetwork via its __init__() method, we introduce two factory class-methods:
* new() - creates a new BFT network locally
* existing() - the BftTestNetwork object acts as a "proxy" for managing the externally provided BFT network.

In the existing() case we also provide the ability to inject externally prepared client instances.

The start_replica() and stop_replica() methods now both support running externally provided commands (useful for managing replicas that run as Docker containers, for example).

Some additional enhancements are required in order to run Apollo as part of product CI (which uses Pytest):
* util becomes a package by now having a __init__.py
* added __test__ = False to test scenarios relevant for product CI (so that PyTest does not directly invoke Apollo tests)